fix: serialize the chromatic bin count as int64_t
JsonWriter overloads on std::int64_t, int, double and bool. On macOS
std::int64_t is long long, so static_cast<long long> bound exactly and the
local build stayed green. On Linux std::int64_t is long, so the argument
matched none of the overloads exactly and every GitHub Actions job — GCC,
Clang and UBSan — failed to compile.

Cast to std::int64_t so the call is exact on both platforms. Local green on one
OS never established cross-platform readiness here.

+++ src/cmd_shading.cpp
@@ -2,6 +2,7 @@
 
 #include <array>
 #include <cmath>
+#include <cstdint>
 #include <cstdlib>
 #include <filesystem>
 #include <iomanip>
@@ -366,7 +367,7 @@ void write_shading_json(std::ostream& os, std::string_view file_label,
   w.key("chromatic_complete");
   w.value(chromatic_available && chroma.complete);
   w.key("missing_chromatic_bin_count");
-  w.value(static_cast<long long>(chroma.missing_bin_count));
+  w.value(static_cast<std::int64_t>(chroma.missing_bin_count));
   w.key("cfa_positions");
   if (chroma.layout_valid || chromatic_available) {
     w.begin_object();
